add Qt6 support and use in static binaries (#529)
- update CMakeLists to work with Qt5.15 / Qt6
- add `-DSME_WITH_QT6=on` bool flag to CMake options
- CI
  - sme_deps -> 2021.05.11 (includes qt6)
  - linux builds: ubuntu 16.04 -> 18.04
  - linux build: add `-DOpenGL_GL_PREFERENCE=LEGACY` to avoid Qt6 picking up unnecessary libOpenGL dependency
  - linux code quality builds: use gcc instead of clang, since they use static libs built with gcc
  - manylinux images -> 2021.05.12
  - python 3.5 is EOL, stop building python 3.5 wheels
- resolves #528

+++ src/core/model/src/model_units.cpp
@@ -110,7 +110,7 @@ static int getOrAddUnitIndex(libsbml::Model *model, const std::string &id,
   // otherwise add imported unit to model & return its index
   units.push_back(std::move(u));
   SPDLOG_INFO("  -> creating new unit '{}'", units.back().name.toStdString());
-  return units.size() - 1;
+  return static_cast<int>(units.size() - 1);
 }
 
 UnitVector::UnitVector(const QVector<Unit> &unitsVec, int defaultIndex)
